reorganize code to prevent coupling
